% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_half.R
\name{calc_half}
\alias{calc_half}
\title{Calculate half the value of x}
\usage{
calc_half(x)
}
\arguments{
\item{x}{an \link{even_number} to take the half of}
}
\value{
an \link{odd_number}
}
\description{
Calculate half the value of x
}
