% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_twice.R
\name{calc_twice}
\alias{calc_twice}
\title{Calculate twice the value of x}
\usage{
calc_twice(x)
}
\arguments{
\item{x}{an \link{even_number} or \link{odd_number} to calculate
twice its value of}
}
\value{
an \link{even_number}
}
\description{
Calculate twice the value of x
}
