% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.even_number}
\alias{print.even_number}
\title{Print an even number}
\usage{
\method{print}{even_number}(even_number)
}
\arguments{
\item{even_number}{an even number of S3 class \link{even_number}}
}
\description{
Print an even number
}
