% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.odd_number}
\alias{print.odd_number}
\title{Print an odd number}
\usage{
\method{print}{odd_number}(value)
}
\arguments{
\item{value}{an odd number of S3 class \link{odd_number}}
}
\description{
Print an odd number
}
