% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{Allow 'print.even_number' to be discovered}
\usage{
print(value, ...)
}
\arguments{
\item{value}{the object to be printed}

\item{...}{whatever}
}
\description{
Allow 'print.even_number' to be discovered
}
